Made 6 maker methods for agent architectures.
